->ajout des yeux \n->correction des parametres \n->optimisation des affichages des constructeurs \n->affichage des visions \n->correction de la géométrie \n->refonte du système de suppression \n->suppression des bestioles fantomes \n->séparation des chemins des clones \n->amélioration de l'affichage \n->amélioration de la capacité des compteurs de bestioles

+++ Camouflages.h
@@ -12,7 +12,7 @@ using namespace std;
 class Camouflages: public Membre
 {
 private:
-   int phi;
+   double phi;
 public :                                        
    Camouflages(DBestiole* dbestiole,double phi);     
    virtual ~Camouflages();
